Simulator Speedup and Slowdown Added
I've added simulator speed up and slow down buttons to adjust the speed the simulation plays at.

I've also adjusted how charge predator works for dodo's. A set of predators  collected when a prey is eaten, storing that in a map with the key of the prey. The set gets updated and grows as new animal types eat others.